modifications

+++ src/profils/Enseignant/views/SuiviStage/Stages/singleEntreprise.js
@@ -1,43 +1,39 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
-import { Link, Redirect,useHistory, useLocation } from 'react-router-dom';
+import { useHistory } from 'react-router-dom';
 
-import {
-    CCard,
-    CCardBody,
-    CCol,
-    CRow
-  } from '@coreui/react'
+import {CCard,CCardBody,CCardFooter,CCol,CRow} from '@coreui/react'
 
 
 
 
 import {Detail, Stage} from './detailEntreprise';
 
-import FormulaireStage
- from './AjouterStage/formAddStage';
+import FormulaireStage from './AjouterStage/formAddStage';
 
- const correspondance_genie = {
-    "Informatique et Télécommunications": "GIT" ,
-    "Génie Civil": "GC" ,
-     "Electro-Mecanique": "GEM",
-     "Aéronautique": "GAR"
-};
 
 
 function Single ({match})  {
 
     let history = useHistory()    
 
+
+    const [reload, setReload] = useState(false)
     useEffect( ()=>{
         refreshLists();
-    },[])
+        getListUser();
+    },[reload])
+
+    
+
+    
+    function refreshPage(e) {
+        setReload(!reload)
+          // window.location.reload(false);
+    }    
+
 
     const [detail, setDetail] = useState({})
-
-
-    
-    
     const refreshLists =  () => {
         axios.get(`http://localhost:8000/api/stage/entreprises/${match.params.slug}/stages/`)
         .then(res => {
@@ -47,15 +43,37 @@ function Single ({match})  {
     }
 
 
-    const deleteEntreprise = (e) =>{
-         axios.delete(`http://localhost:8000/api/stage/entreprises/${match.params.slug}/delete/`)
-          .then(res => {
-            console.log('supprimé');
-            alert('SUPPRESSION REUSSIE')
-            return history.push("/enseignant/entreprises");
+    const [cbon, setCbon] = useState(false)
+    const [user, setUser] = useState([])
+    const getListUser = () =>{
+        axios.get(`http://users-ent.herokuapp.com/api/auth/ETUDIANT/`)
+        .then((res)=>{
+            //console.log(res.data)
+            setUser(res.data)
+            setCbon(true);
         })
-        .catch(err => console.log(err))
+        .catch(e=>console.log(e))
+    }
+
+
+
+    const deleteEntreprise = (e) =>{
+
+        let confirmer = window.confirm("Voulez-vous vraiment supprimer cette entreprise ?");
+        if(confirmer){
+            axios.delete(`http://localhost:8000/api/stage/entreprises/${match.params.slug}/delete/`)
+            .then(() => {
+                console.log('supprimé');
+                alert('SUPPRESSION REUSSIE')
+                return history.push("/enseignant/entreprises");
+            })
+            .catch(err => console.log(err))
+        }
+        else{
+            return null;
+        }         
     }  
+
 
 
 
@@ -65,13 +83,14 @@ function Single ({match})  {
         <CCard>
         <CCardBody>
         <CRow>
+            {(detail && cbon) &&
             <CCol sm="12">
                 <Detail
                 nom_entreprise={detail.nom_entreprise}
-                domaine = {detail.domaine}
-                classe_stage={detail.classe_stage}
+                type_stage = {detail.type_stage}
                 adresse = {detail.adresse_entreprise}
                 telephone_entreprise={detail.telephone_entreprise}
+                slug={match.params.slug}
                 supprimer = {e => deleteEntreprise()}
                 
                 />
@@ -82,16 +101,22 @@ function Single ({match})  {
                 {detail.stages ?
                     <CRow style={{marginTop:'2em'}}>
                     
-                    {detail['stages'].map((stage) =>
+                    {detail['stages'].map((stage) =>{
+                        let eleve = (user.filter(el=> el.id == stage.stagiaire))[0]
                         
-                        <Stage
-                            nom={stage.stagiaire? stage.stagiaire.etudiant.username: 'Non alloué' }
-                            classe={stage.stagiaire && stage.stagiaire.classe}
-                            genie = {stage.stagiaire && stage.stagiaire.genie}
+                    return <Stage
+                            nom={stage.stagiaire && eleve.user.last_name}
+                            prenom = {stage.stagiaire && eleve.user.first_name}
+                            classe={stage.stagiaire && eleve.classe }
+                            genie = {stage.stagiaire && eleve.departement}
                             numero = {stage.id}
+                            slug={match.params.slug}
+                            debut_stage = {stage.date_debut_stage}
+                            fin_stage={stage.date_fin_stage}
                         />
-                         
-                    )}
+                    }
+
+                )}
                    
                     </CRow>
                     :
@@ -99,19 +124,20 @@ function Single ({match})  {
                 }
                 </>
             </CCol>
+            }
         </CRow>
         </CCardBody>
 
+        <CCardFooter>
 
-        <div>
         <FormulaireStage
-            classe={detail.classe_stage}
-            genie = {correspondance_genie[`${detail.domaine}`]}
             slug={match.params.slug}
             entreprise={detail.nom_entreprise}
+            clickLoad={e=>refreshPage()}
         />
-        </div>
-        
+
+        </CCardFooter>
+
         </CCard>
     )
 }
